fix: storybook stories [MONET-1338]
