+++ packages/ecommerce-app-base/.storybook/parameters.tsx
@@ -1,0 +1,12 @@
+import { StoryFn } from '@storybook/react';
+
+export const parameters = {
+  tags: ['autodocs'],
+  decorators: [
+    (Story: StoryFn) => (
+      <div style={{ width: '700px' }}>
+        <Story />
+      </div>
+    ),
+  ],
+};
