+++ packages/ecommerce-app-base/.storybook/preview.ts
@@ -2,6 +2,7 @@ import type { Preview } from '@storybook/react';
 
 const preview: Preview = {
   parameters: {
+    layout: 'centered',
     actions: { argTypesRegex: '^on[A-Z].*' },
     controls: {
       matchers: {
