+++ packages/ecommerce-app-base/src/Editor/Field.stories.tsx
@@ -8,10 +8,12 @@ import { IntegrationProvider } from './IntegrationContext';
 import { sdk } from '../__mocks__/storybook/sdk';
 import { GlobalStyles } from '@contentful/f36-components';
 import { Integration } from '../interfaces';
+import { parameters } from '../../.storybook/parameters';
 
 const meta: Meta<typeof Field> = {
   component: Field,
   tags: ['autodocs'],
+  ...parameters,
 };
 
 export default meta;
