+++ packages/ecommerce-app-base/src/ProductCard/context/index.tsx
@@ -1,0 +1,5 @@
+export {
+  ResourceFieldContext,
+  useResourceField,
+  ResourceFieldProvider,
+} from './ResourceFieldContext';
