+++ packages/ecommerce-app-base/src/SkuPicker/ProductSelectionList/ProductSelectionList.stories.tsx
@@ -12,9 +12,29 @@ export default meta;
 
 type Story = StoryObj<typeof ProductSelectionList>;
 
+const productPreviewNoImages = productPreviews.map((value) => {
+  const product = { ...value, image: '' };
+  return product;
+});
+
 export const Default: Story = {
   args: {
     products: productPreviews,
+    selectProduct: (...args) => alert(args),
+  },
+  decorators: [
+    (Story) => (
+      <>
+        <GlobalStyles />
+        <Story />
+      </>
+    ),
+  ],
+};
+
+export const NoImages: Story = {
+  args: {
+    products: productPreviewNoImages,
     selectProduct: (...args) => alert(args),
   },
   decorators: [
