+++ packages/ecommerce-app-base/src/__mocks__/productPreviews.ts
@@ -3,21 +3,24 @@ import { Product } from '../interfaces';
 const productPreviews: Product[] = [
   {
     id: '28a7442d',
-    image: 'https://example.com/commercetools-example/products/picture-1.jpg',
+    image:
+      'https://images.ctfassets.net/juh8bvgveao4/QoAO8rqn86a4jiH1yudiN/e518fd9263b67705c3ffb041bd217bda/imageService.webp',
     name: 'Dress Twin-Set rose',
     sku: 'M0E20130820E90Z',
     externalLink: 'https://mc.commercetools.com/example-project/products/28a7442d/general',
   },
   {
     id: '4e5446e8',
-    image: 'https://example.com/commercetools-example/products/picture-2.jpg',
+    image:
+      'https://images.ctfassets.net/juh8bvgveao4/QoAO8rqn86a4jiH1yudiN/e518fd9263b67705c3ffb041bd217bda/imageService.webp',
     name: 'Michael Kors – Shopper “Jet Set Travel”',
     sku: 'A0E2300FX102203',
     externalLink: 'https://mc.commercetools.com/example-project/products/4e5446e8/general',
   },
   {
     id: 'db3c234l',
-    image: 'https://example.com/commercetools-example/products/picture-3.jpg',
+    image:
+      'https://images.ctfassets.net/juh8bvgveao4/QoAO8rqn86a4jiH1yudiN/e518fd9263b67705c3ffb041bd217bda/imageService.webp',
     name: 'Shirt “Jenny“ Polo Ralph Lauren white',
     sku: 'M0E21300900DZN7',
     externalLink: 'https://mc.commercetools.com/example-project/products/db3c234l/general',
