+++ packages/ecommerce-app-base/tsconfig.json
@@ -15,7 +15,7 @@
       "es2017"
     ],
     "resolveJsonModule": true,
-    "declaration": true
+    "declaration": true,
   },
   "files": [
     "./src/index.tsx"
